average spike display

+++ src/engine/SpikeAnalysis.h
@@ -50,7 +50,7 @@ public:
 	// filename, chan: file and channel to use
 	
 	static const double SPIKEHALFLEN = 0.002; // s
-	static void averageWaveform(std::vector<double> &average, std::vector<double> &std,
+	static void averageWaveform(std::vector<float> &average, std::vector<float> &std,
 			const std::vector<int64_t> &train, const char *filename, int chan); 
 };
